requirements.txt
fastapi
uvicorn
SkPy
python-dotenv